Reorganized the index.ts and added authRoutes.ts
